Session_5/mailer/pyproject.toml
[project]
name = "mailer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []

[tool.poetry.dependencies]
requests = "*"
beautifulsoup4 = "*"
python-docx = "*"
rich = "*"
